Replace initializer and check for header & footer
Fixes #7

Overriding the default container.lookup to avoid errors when a consuming
app does not have a header or footer template seems drastic.

The method in this commit instead checks the container for the templates
and only returns the template path to the partial helpers if it is found.

However, in Ember 1.11.0, `this.container.has` is deprecated.
There may be another way to go about this altogether, but for now this
seems to work until a better solution is found.